Move Ty accessors to TyExt
